Added controllers and error handling

+++ backend/src/main/java/com/bootcamp/backend/backend/employee/EmployeeService.java
@@ -14,26 +14,51 @@ public class EmployeeService {
     }
 
     public Employee addEmployee(Employee employee) {
+        if (employee.getId() != null && employeeRepository.existsById(employee.getId())) {
+            throw new EmployeeAlreadyExistsException();
+        }
+
         return employeeRepository.save(employee);
     }
 
     public void deleteEmployeeById(String id) {
+        Optional<Employee> foundEmployee = employeeRepository.findById(id);
+        if (foundEmployee.isEmpty()) {
+            throw new EmployeeNotFoundException();
+        }
         employeeRepository.deleteById(id);
     }
 
-    public Optional<Employee> getEmployeeById(String id) {
-        return employeeRepository.findById(id);
+    public Employee getEmployeeById(String id) {
+        Optional<Employee> foundEmployee = employeeRepository.findById(id);
+        if (foundEmployee.isEmpty()) {
+            throw new EmployeeNotFoundException();
+        }
+
+        return foundEmployee.get();
     }
 
     public List<Employee> getEmployees() {
         return employeeRepository.findAll();
     }
 
-    public List<Employee> getEmployeesByNameOrSurnameContaining(String term) {
-        return employeeRepository.findByNameOrSurnameContainingIgnoreCase(term);
+    public List<Employee> getEmployeesByNameOrSurnameStartingWith(String term) {
+        List<Employee> employeeList = employeeRepository.findByNameOrSurnameStartingWithIgnoreCase(term);
+        if (employeeList.isEmpty()) {
+            throw new EmployeeNotFoundException();
+        }
+
+        return employeeList;
     }
 
-    public Employee updateEmployee(Employee employee) {
+    public Employee updateEmployee(String idFromPath, Employee employee) {
+        if (!employee.getId().equals(idFromPath)) {
+            throw new DifferentEmployeeIdInDatabaseException();
+        }
+        if (!employeeRepository.existsById(employee.getId())) {
+            throw new EmployeeNotFoundException();
+        }
+
         return employeeRepository.save(employee);
     }
 }
