+++ backend/src/main/java/com/bootcamp/backend/backend/skill/SkillService.java
@@ -14,14 +14,27 @@ public class SkillService {
     }
 
     public Skill addSkill(Skill skill) {
+        if (skill.getId() != null && skillRepository.existsById(skill.getId())) {
+            throw new SkillAlreadyExistsException();
+        }
+
         return skillRepository.save(skill);
     }
 
     public void deleteSkillById(Long id) {
+        Optional<Skill> foundSkill = skillRepository.findById(id);
+        if (foundSkill.isEmpty()) {
+            throw new SkillNotFoundException();
+        }
         skillRepository.deleteById(id);
     }
 
     public Optional<Skill> getSkillById(Long id) {
+        Optional<Skill> foundSkill = skillRepository.findById(id);
+        if (foundSkill.isEmpty()) {
+            throw new SkillNotFoundException();
+        }
+
         return skillRepository.findById(id);
     }
 
@@ -33,7 +46,14 @@ public class SkillService {
         return skillRepository.findByNameContainingIgnoreCase(term);
     }
 
-    public Skill updateSkill(Skill skill) {
+    public Skill updateSkill(Long idFromPath, Skill skill) {
+        if (!skill.getId().equals(idFromPath)) {
+            throw new DifferentSkillIdInDatabaseException();
+        }
+        if (!skillRepository.existsById(skill.getId())) {
+            throw new SkillNotFoundException();
+        }
+
         return skillRepository.save(skill);
     }
 }
